refactor: economy phase-2 slice 8 — D5 and validator scenario tables (#3856) (#3864)
* refactor: economy phase-2 slice 8 — D5 and validator scenario tables (#3856)

Migrate first_right_of_refusal_issue_acceptance_criteria_d5_test and
world_market_trade_order_validator_test to shared scenario tables in
colonizethis_economy_test_support; D5 test drops colonizethis_logic import.

Refs #3856


* refactor: economy phase-2 slice 9 — first-right profit scenario tables (#3856)

Hoist inline const tables from first_right_profit_test.dart into
colonizethis_economy_test_support with run* helpers; preserve all
test description pins for duplicate-description CI.


* refactor: economy phase-2 slice 10 — consumption scenario tables (#3856)

Migrate resolveConsumption and per-phase consumption helper tests to
table-driven scenarios in colonizethis_economy_test_support.


* fix: split resolveConsumptionScenarios to satisfy function-size gate

Refs #3856


* refactor: economy phase-2 slice 11 — build-cost scenario tables (#3856)

Migrate build_cost_test.dart to table-driven scenarios in test_support,
preserving all nine test descriptions for duplicate-description CI.


* refactor: economy phase-2 slice 12 — production scenario tables (#3856)

Migrate economy_production_test.dart to table-driven scenarios in
colonizethis_economy_test_support; preserve all eleven test descriptions.

Refs #3856


* refactor: economy phase-2 slice 13 — worker action cost scenario tables (#3856)

Hoist recruit/train affordability and deduction rows into
worker_action_cost_scenarios.dart; runner uses table loops.


* refactor: economy phase-2 slice 14 — non-GP auto-offers scenario tables (#3856)

Migrate the last core economy imperative suite to table-driven scenarios
in test_support, completing phase-2 scenario consolidation for #3856.


---------
