+++ packages/colonizethis_economy/test/economy/world_market/first_right_profit_test.dart
@@ -1,6 +1,8 @@
-import 'package:colonizethis_test/test.dart';
-import 'package:colonizethis_economy/colonizethis_economy.dart';
+// Table-driven unit tests for FirstRightProfit (Refs #3856).
 
+import 'package:colonizethis_economy/colonizethis_economy.dart';
+import 'package:colonizethis_economy_test_support/colonizethis_economy_test_support.dart';
+import 'package:colonizethis_test/test.dart';
 
 void main() {
   group('FirstRightProfit constants', () {
@@ -23,22 +25,9 @@ void main() {
   });
 
   group('computeFirstRightProfitRate (#3753 R8.2 — full relation-linear)', () {
-    const rateCases = <({int score, double expected})>[
-      (score: 0, expected: 0.0),
-      (score: 75, expected: 0.75),
-      (score: 100, expected: kFirstRightMaxProfitRate),
-      (score: -25, expected: 0.0),
-      (score: 150, expected: kFirstRightMaxProfitRate),
-    ];
-
-    for (final c in rateCases) {
-      test('returns ${c.expected} at relationScore ${c.score}', () {
-        final result = computeFirstRightProfitRate(c.score);
-        if (c.expected == kFirstRightMaxProfitRate || c.expected == 0.0) {
-          expect(result, c.expected);
-        } else {
-          expect(result, closeTo(c.expected, 1e-12));
-        }
+    for (final scenario in firstRightProfitRateScenarios) {
+      test(scenario.label, () {
+        runFirstRightProfitRateScenario(scenario);
       });
     }
 
@@ -54,100 +43,10 @@ void main() {
   });
 
   group('computeFirstRightProfit (#2992 D3 + D4 helper)', () {
-    const profitCases = <({
-      int relationScore,
-      int filledQuantity,
-      double pricePerUnit,
-      double expectedRate,
-      double expectedTreasury,
-      bool expectZero,
-    })>[
-      (
-        relationScore: 0,
-        filledQuantity: 10,
-        pricePerUnit: 5.0,
-        expectedRate: 0.0,
-        expectedTreasury: 0.0,
-        expectZero: true,
-      ),
-      (
-        relationScore: 75,
-        filledQuantity: 10,
-        pricePerUnit: 5.0,
-        expectedRate: 0.75,
-        expectedTreasury: 37.5,
-        expectZero: false,
-      ),
-      (
-        relationScore: 100,
-        filledQuantity: 1,
-        pricePerUnit: 1.0,
-        expectedRate: kFirstRightMaxProfitRate,
-        expectedTreasury: 1.0,
-        expectZero: false,
-      ),
-      (
-        relationScore: 100,
-        filledQuantity: 4,
-        pricePerUnit: 2.5,
-        expectedRate: kFirstRightMaxProfitRate,
-        expectedTreasury: 10.0,
-        expectZero: false,
-      ),
-      (
-        relationScore: 100,
-        filledQuantity: 0,
-        pricePerUnit: 5.0,
-        expectedRate: 0.0,
-        expectedTreasury: 0.0,
-        expectZero: true,
-      ),
-      (
-        relationScore: 100,
-        filledQuantity: 10,
-        pricePerUnit: 0.0,
-        expectedRate: 0.0,
-        expectedTreasury: 0.0,
-        expectZero: true,
-      ),
-      (
-        relationScore: 100,
-        filledQuantity: -5,
-        pricePerUnit: 5.0,
-        expectedRate: 0.0,
-        expectedTreasury: 0.0,
-        expectZero: true,
-      ),
-      (
-        relationScore: 100,
-        filledQuantity: 5,
-        pricePerUnit: -1.0,
-        expectedRate: 0.0,
-        expectedTreasury: 0.0,
-        expectZero: true,
-      ),
-    ];
-
-    for (final c in profitCases) {
-      test(
-        'relationScore ${c.relationScore}, qty ${c.filledQuantity} @ '
-        '${c.pricePerUnit} → rate ${c.expectedRate}, treasury '
-        '${c.expectedTreasury}',
-        () {
-          final result = computeFirstRightProfit(
-            relationScore: c.relationScore,
-            filledQuantity: c.filledQuantity,
-            pricePerUnit: c.pricePerUnit,
-          );
-          if (c.expectZero) {
-            expect(result, FirstRightProfit.zero);
-            expect(result.profitTreasury, 0.0);
-          } else {
-            expect(result.profitRate, closeTo(c.expectedRate, 1e-12));
-            expect(result.profitTreasury, closeTo(c.expectedTreasury, 1e-12));
-          }
-        },
-      );
+    for (final scenario in firstRightProfitScenarios) {
+      test(scenario.label, () {
+        runFirstRightProfitScenario(scenario);
+      });
     }
 
     test('treasury equals filledQuantity * pricePerUnit * profitRate', () {
@@ -181,36 +80,10 @@ void main() {
   });
 
   group('computeEmbassyKickback (#3753 R8.3 — 10% of relation portion)', () {
-    const kickbackCases = <({
-      num relationScore,
-      int filledQuantity,
-      double pricePerUnit,
-      double expected,
-    })>[
-      (relationScore: 100, filledQuantity: 10, pricePerUnit: 20.0, expected: 20.0),
-      (relationScore: 50, filledQuantity: 10, pricePerUnit: 20.0, expected: 10.0),
-      (relationScore: 80.0, filledQuantity: 4, pricePerUnit: 2.5, expected: 0.8),
-      (relationScore: 0, filledQuantity: 10, pricePerUnit: 20.0, expected: 0.0),
-      (relationScore: 100, filledQuantity: -5, pricePerUnit: 20.0, expected: 0.0),
-      (relationScore: 100, filledQuantity: 10, pricePerUnit: 0.0, expected: 0.0),
-      (relationScore: 150, filledQuantity: 10, pricePerUnit: 20.0, expected: 20.0),
-    ];
-
-    for (final c in kickbackCases) {
-      test(
-        'relation ${c.relationScore}, ${c.filledQuantity} @ ${c.pricePerUnit} '
-        '→ ${c.expected}',
-        () {
-          expect(
-            computeEmbassyKickback(
-              relationScore: c.relationScore,
-              filledQuantity: c.filledQuantity,
-              pricePerUnit: c.pricePerUnit,
-            ),
-            closeTo(c.expected, 1e-12),
-          );
-        },
-      );
+    for (final scenario in embassyKickbackScenarios) {
+      test(scenario.label, () {
+        runEmbassyKickbackScenario(scenario);
+      });
     }
   });
 }
