+++ packages/colonizethis_economy_test_support/lib/colonizethis_economy_test_support.dart
@@ -8,11 +8,15 @@
 library colonizethis_economy_test_support;
 
 export 'src/bid_spend_game_factory.dart';
+export 'src/build_cost_scenarios.dart';
 export 'src/bid_spend_parity_scenarios.dart';
 export 'src/boycott_blocked_commodities_scenarios.dart';
 export 'src/boycott_blocked_commodities_test_support.dart';
 export 'src/carry_forward_bid_notional_scenarios.dart';
+export 'src/consumption_phases_scenarios.dart';
+export 'src/consumption_scenarios.dart';
 export 'src/core_economy_test_support.dart';
+export 'src/economy_production_scenarios.dart';
 export 'src/deal_matcher_frr_scenarios.dart';
 export 'src/deal_matcher_priority_scenarios.dart';
 export 'src/deal_matcher_scenarios.dart';
@@ -22,10 +26,13 @@ export 'src/deal_matcher_treasury_scenarios.dart';
 export 'src/effective_market_price_scenarios.dart';
 export 'src/frr_credits_scenarios.dart';
 export 'src/frr_credits_test_support.dart';
+export 'src/frr_issue_ac_d5_scenarios.dart';
+export 'src/first_right_profit_scenarios.dart';
 export 'src/gp_treasury_credit_accumulator_scenarios.dart';
 export 'src/lock_recovery_minor_bids_scenarios.dart';
 export 'src/lock_recovery_minor_bids_test_support.dart';
 export 'src/matcher_treasury_budget_scenarios.dart';
+export 'src/non_gp_auto_offers_scenarios.dart';
 export 'src/non_gp_auto_offers_test_support.dart';
 export 'src/non_gp_extraction_scenarios.dart';
 export 'src/non_gp_extraction_test_support.dart';
@@ -43,6 +50,7 @@ export 'src/purchased_tile_riches_scenarios.dart';
 export 'src/purchased_tile_riches_test_support.dart';
 export 'src/treasury_bid_budget_test_support.dart';
 export 'src/treasury_bid_budget_scenarios.dart';
+export 'src/worker_action_cost_scenarios.dart';
 export 'src/trade_order_admission_scenarios.dart';
 export 'src/trade_order_factory.dart';
 export 'src/trade_order_suggester_scenarios.dart';
